refactor(mobile): partir nueva_cita_form.dart en 4 modulos (fase 5a-1)
- nueva_cita_form.dart  320  pantalla + state (fetch, navegacion mes)
- calendario_mes.dart   251  calendario, celdas dia/header, leyenda
- slots_horarios.dart   171  cabecera, grid, chip + constantes horario
- confirmacion_sheet.dart 190 bottom sheet confirmar

Archivo original (914 lineas) movido a subcarpeta nueva_cita/.
Sin cambios funcionales. flutter analyze: 0 issues.
